frontend additions

+++ ingestion-server/config/redis.go
@@ -11,11 +11,20 @@ import (
 type LogCache struct {
 	NextPageState string
 	Logs          []types.Log
+	Total         int
 }
 
 var ctx = context.TODO()
 
+func isCachingEnabled() bool {
+	return initializers.CONFIG.CACHING
+}
+
 func GetCache(key string) LogCache {
+	if !isCachingEnabled() {
+		return LogCache{}
+	}
+
 	data, err := initializers.RedisClient.Get(ctx, key).Result()
 	if err != nil {
 		fmt.Println(err)
@@ -33,6 +42,10 @@ func GetCache(key string) LogCache {
 }
 
 func GetCacheKeys() []string {
+	if !isCachingEnabled() {
+		return nil
+	}
+
 	keys, err := initializers.RedisClient.Keys(ctx, "*").Result()
 	if err != nil {
 		Logger.Warnw("Failed to get cache keys", "Error", err)
@@ -43,6 +56,10 @@ func GetCacheKeys() []string {
 }
 
 func SetCache(key string, logs LogCache) {
+	if !isCachingEnabled() {
+		return
+	}
+
 	data, err := json.Marshal(logs)
 	if err != nil {
 		Logger.Warnw("Failed to marshal logs", "Error", err)
@@ -53,6 +70,10 @@ func SetCache(key string, logs LogCache) {
 }
 
 func RemoveCache(key string) {
+	if !isCachingEnabled() {
+		return
+	}
+
 	err := initializers.RedisClient.Del(ctx, key).Err()
 	if err != nil {
 		Logger.Warnw("Failed to remove cache", "Error", err)
@@ -60,6 +81,10 @@ func RemoveCache(key string) {
 }
 
 func FlushCache() {
+	if !isCachingEnabled() {
+		return
+	}
+
 	err := initializers.RedisClient.FlushAll(ctx).Err()
 	if err != nil {
 		Logger.Warnw("Failed to flush cache", "Error", err)
